+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/actions/impl/ActionHelperImpl.java
@@ -22,6 +22,7 @@ package com.adobe.aem.commons.assetshare.components.actions.impl;
 import com.adobe.aem.commons.assetshare.components.actions.ActionHelper;
 import com.adobe.aem.commons.assetshare.configuration.Config;
 import com.adobe.aem.commons.assetshare.content.AssetModel;
+import com.adobe.granite.asset.api.AssetException;
 import com.day.cq.wcm.api.WCMMode;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.request.RequestParameter;
@@ -30,6 +31,10 @@ import org.apache.sling.models.factory.ModelFactory;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
+import javax.net.ssl.StandardConstants;
+import java.io.UnsupportedEncodingException;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Collection;
 
@@ -45,14 +50,22 @@ public final class ActionHelperImpl implements ActionHelper {
 
         if (requestParameters != null) {
             for (final RequestParameter requestParameter : requestParameters) {
-                final Resource resource = request.getResourceResolver().getResource(requestParameter.getString());
+                String path = requestParameter.getString();
+
+                try {
+                    path = URLDecoder.decode(path, StandardCharsets.UTF_8.name());
+                } catch (UnsupportedEncodingException ex) {
+                    throw new AssetException("Could not UTF-8 encode the asset path.", requestParameter.getString());
+                }
+
+                final Resource resource = request.getResourceResolver().getResource(path);
                 if (resource != null) {
                     final AssetModel asset = modelFactory.getModelFromWrappedRequest(request, resource, AssetModel.class);
-
                     if (asset != null) {
                         assets.add(asset);
                     }
                 }
+
             }
         }