+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/actions/share/impl/EmailShareServiceImpl.java
@@ -40,6 +40,7 @@ import com.adobe.granite.security.user.UserProperties;
 import com.adobe.granite.security.user.UserPropertiesManager;
 import com.day.cq.commons.Externalizer;
 import com.day.cq.dam.commons.util.DamUtil;
+import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.jackrabbit.api.security.user.Authorizable;
 import org.apache.sling.api.SlingHttpServletRequest;
@@ -61,10 +62,11 @@ import org.osgi.service.metatype.annotations.ObjectClassDefinition;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import javax.jcr.RepositoryException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+
+import java.io.UnsupportedEncodingException;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
+import java.util.*;
 
 @Component(service = ShareService.class)
 @Designate(ocd = EmailShareServiceImpl.Cfg.class)
@@ -110,7 +112,7 @@ public class EmailShareServiceImpl implements ShareService {
     @Override
     public final void share(final SlingHttpServletRequest request, final SlingHttpServletResponse response, final ValueMap shareParameters) throws ShareException {
     	
-    		/** Work around for regression issue introduced in AEM 6.4 **/
+        /** Work around for regression issue introduced in AEM 6.4 **/
         SlingBindings bindings = new SlingBindings();
         //intentionally setting the second argument to 'null' since there is no SlingScript to pass in
         bindings.setSling( new ScriptHelper(bundleContext, null, request, response));
@@ -120,7 +122,7 @@ public class EmailShareServiceImpl implements ShareService {
 
         shareParameters.putAll(xssProtectUserData(emailShare.getUserData()));
 
-        // Configured data supercedes user data
+        // Configured data supersedes user data
         shareParameters.putAll(emailShare.getConfiguredData());
 
         // Except for signature which we may or may  not want to use from configured data, depending on flags in configured data
@@ -135,15 +137,24 @@ public class EmailShareServiceImpl implements ShareService {
 
     private final void share(final Config config, final ValueMap shareParameters, final String emailTemplatePath) throws ShareException {
         final String[] emailAddresses = StringUtils.split(shareParameters.get(EMAIL_ADDRESSES, ""), ",");
-        final String[] assetPaths = shareParameters.get(ASSET_PATHS, String[].class);
+        final String[] assetPaths = Arrays.stream(shareParameters.get(ASSET_PATHS, ArrayUtils.EMPTY_STRING_ARRAY))
+                .map(path -> {
+                    try {
+                        return URLDecoder.decode(path, StandardCharsets.UTF_8.name());
+                    } catch (UnsupportedEncodingException e) {
+                        log.warn("Could not decode path [ {} ] as UTF-8; Using path as is,.", path);
+                        return path;
+                    }
+                }).filter(StringUtils::isNotBlank)
+                .toArray(String[]::new);
 
         // Check to ensure the minimum set of e-mail parameters are provided; Throw exception if not.
         if (emailAddresses == null || emailAddresses.length == 0) {
             throw new ShareException("At least one e-mail address is required to share");
-        } else if (assetPaths == null || assetPaths.length == 0) {
+        } else if (ArrayUtils.isEmpty(assetPaths)) {
             throw new ShareException("At least one asset is required to share");
         }
-
+       
         // Convert provided params to <String, String>; anything that needs to be accessed in its native type should be accessed and manipulated via shareParameters.get(..)
         final Map<String, String> emailParameters = new HashMap<String, String>();
         for (final String key : shareParameters.keySet()) {
@@ -226,7 +237,7 @@ public class EmailShareServiceImpl implements ShareService {
     /**
      * Since all emails are expected to be HTML emails in this implementation, we must XSS Protect all values that may end up in the HTML email.
      *
-     * @param userData the Map of data provided by the end-user (usually derived from the Request) to use in the email.
+     * @param dirtyUserData the Map of data provided by the end-user (usually derived from the Request) to use in the email.
      * @return the protected Map; all String's are xss protected for HTML.
      */
     private Map<String, Object> xssProtectUserData(Map<String, Object> dirtyUserData) {