+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/details/Image.java
@@ -23,4 +23,6 @@ public interface Image extends EmptyTextComponent {
     String getSrc();
 
     String getAlt();
+
+    default String getFallback() { return null; }
 }