+++ core/src/main/java/com/adobe/aem/commons/assetshare/util/MimeTypeHelper.java
@@ -1,0 +1,13 @@
+package com.adobe.aem.commons.assetshare.util;
+
+import org.osgi.annotation.versioning.ProviderType;
+
+@ProviderType
+public interface MimeTypeHelper {
+    /**
+     * @param mimeType the mime type to check
+     * @return true if the mime type is to be considered supported by the browser.
+     */
+    boolean isBrowserSupportedImage(String mimeType);
+
+}